chore: fix layout && change email btn to whastApp
