+++ src/components/MainLayout/FixedUI.tsx
@@ -1,5 +1,4 @@
 import { Toaster } from "sonner";
-import WhatsAppButton from "../WhatsAppButton";
 
 const FixedUI = () => {
   return (
@@ -10,11 +9,6 @@ const FixedUI = () => {
         expand={true}
         closeButton
       />
-
-      {/* Fixed WhatsApp Button */}
-      <div className='fixed right-6 bottom-6 z-50'>
-        <WhatsAppButton />
-      </div>
     </>
   );
 };
