+++ src/components/MainLayout/Footer.tsx
@@ -1,3 +1,4 @@
+import posthog from "posthog-js";
 import { Link } from "react-router-dom";
 import Github from "../../assets/icons/github.svg";
 import Linkedin from "../../assets/icons/linkedIn.svg";
@@ -5,13 +6,26 @@ import X from "../../assets/icons/x.svg";
 import { Button } from "../ui/button";
 
 const Footer = () => {
+  const phoneNumber = import.meta.env.VITE_WHATSAPP_NUMBER;
+
+  const message = encodeURIComponent(
+    `explain the problem and, if possible, add a screenshot.
+    .اذا في اي مشكلة ارفق معاه سكرينشوت`
+  );
+
+  const handleClick = () => {
+    const url = `https://wa.me/${phoneNumber}?text=${message}`;
+    const newWindow = window.open(url, "_blank", "noopener,noreferrer");
+    if (!newWindow) {
+      posthog.capture("whatsapp");
+      console.warn("Failed to open WhatsApp - popup may be blocked");
+    }
+  };
   return (
-    <footer className='flex flex-col gap-4 py-6 text-black shadow-[0_-4px_12px_rgba(0,0,0,0.1)] sm:flex-row sm:justify-center dark:border-neutral-800 dark:text-white'>
+    <footer className='flex flex-col gap-3 py-3 text-black shadow-[0_-4px_12px_rgba(0,0,0,0.1)] sm:flex-row sm:justify-center sm:py-6 dark:border-neutral-800 dark:text-white'>
       {/* Contact & Policy */}
       <div className='flex justify-center gap-3'>
-        <Button asChild>
-          <a href='mailto:mohammedjrt@gmail.com'>Contact me via email</a>
-        </Button>
+        <Button onClick={handleClick}>Contact me via WhatsApp</Button>
         <Button asChild>
           <Link to='/privacy-policy'>Policy & Terms</Link>
         </Button>
