+++ src/components/WhatsAppButton.tsx
@@ -1,33 +1,34 @@
-import whatsAppIcon from "@/assets/icons/whatsApp.webp";
-import posthog from "posthog-js";
-const WhatsAppButton = () => {
-  const phoneNumber = import.meta.env.VITE_WHATSAPP_NUMBER;
+// import whatsAppIcon from "@/assets/icons/whatsApp.webp";
+// import posthog from "posthog-js";
 
-  const message = encodeURIComponent(
-    "explain the problem and, if possible, add a screenshot"
-  );
+// const WhatsAppButton = () => {
+// const phoneNumber = import.meta.env.VITE_WHATSAPP_NUMBER;
 
-  const handleClick = () => {
-    const url = `https://wa.me/${phoneNumber}?text=${message}`;
-    const newWindow = window.open(url, "_blank", "noopener,noreferrer");
-    if (!newWindow) {
-      posthog.capture("whatsapp");
-      console.warn("Failed to open WhatsApp - popup may be blocked");
-    }
-  };
+// const message = encodeURIComponent(
+//   "explain the problem and, if possible, add a screenshot"
+// );
 
-  return (
-    <button
-      onClick={handleClick}
-      className='fixed right-6 bottom-6 z-50 flex size-10 items-center justify-center rounded-full shadow-lg transition-colors'
-      aria-label='Chat on WhatsApp'>
-      <img
-        src={whatsAppIcon}
-        className='size-10'
-        alt='whatsapp logo'
-      />
-    </button>
-  );
-};
+// const handleClick = () => {
+//   const url = `https://wa.me/${phoneNumber}?text=${message}`;
+//   const newWindow = window.open(url, "_blank", "noopener,noreferrer");
+//   if (!newWindow) {
+//     posthog.capture("whatsapp");
+//     console.warn("Failed to open WhatsApp - popup may be blocked");
+//   }
+// };
 
-export default WhatsAppButton;
+//   return (
+//     <button
+//       onClick={handleClick}
+//       className={`sticky bottom-0 z-50 w-full overflow-x-visible rounded-full p-0 shadow-lg`}
+//       aria-label='Chat on WhatsApp'>
+//       <img
+//         src={whatsAppIcon}
+//         className='absolute -top-20 right-4 h-16 w-16'
+//         alt='whatsapp logo'
+//       />
+//     </button>
+//   );
+// };
+
+// export default WhatsAppButton;
