+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/constants.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/DevAlert.tsx","./src/components/ErrorBoundry.tsx","./src/components/ErrorElement.tsx","./src/components/FormFields.tsx","./src/components/Hero.tsx","./src/components/HydrateFallBack.tsx","./src/components/NotFound.tsx","./src/components/PrivateRoute.tsx","./src/components/ScrollToTopButton.tsx","./src/components/SpinnerOverlay.tsx","./src/components/MainLayout/Footer.tsx","./src/components/MainLayout/Header.tsx","./src/components/MainLayout/Logo.tsx","./src/components/MainLayout/MobileAuthBtns.tsx","./src/components/MainLayout/MobileMenu.tsx","./src/components/MainLayout/ProfilePic.tsx","./src/components/MainLayout/SigninAndLogoutButton.tsx","./src/components/MainLayout/app-sidebar.tsx","./src/components/MainLayout/mode-toggle.tsx","./src/components/auth/FaceBookOAuth.tsx","./src/components/auth/GoogleOAuth.tsx","./src/components/drugInfo/DrugInfoAccordion.tsx","./src/components/drugInfo/DrugPropertyDescription.tsx","./src/components/drugInfo/SearchDrugInfo.tsx","./src/components/drugInfo/drugCard.tsx","./src/components/drugList/InfinitScroll.tsx","./src/components/drugList/ListItem.tsx","./src/components/drugList/SelectRoute.tsx","./src/components/tellUsMore/SelectWithOther.tsx","./src/components/ui/accordion.tsx","./src/components/ui/alert.tsx","./src/components/ui/autocomplete.tsx","./src/components/ui/avatar.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/command.tsx","./src/components/ui/dialog.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/popover.tsx","./src/components/ui/select.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/sidebar.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/sonner.tsx","./src/components/ui/tooltip.tsx","./src/hooks/authContext.tsx","./src/hooks/use-mobile.ts","./src/hooks/useAuth.tsx","./src/hooks/useGetUser.ts","./src/hooks/useInfiniteScroll.ts","./src/hooks/useTheme.ts","./src/layouts/MainLayout.tsx","./src/lib/api.ts","./src/lib/firebase.ts","./src/lib/firebaseAuth.ts","./src/lib/firestore.ts","./src/lib/indexedDB.ts","./src/lib/queryQlient.ts","./src/lib/router.tsx","./src/lib/schemas.ts","./src/lib/utils.ts","./src/pages/DrugInfo.tsx","./src/pages/DrugList.tsx","./src/pages/GlobalError.tsx","./src/pages/Home.tsx","./src/pages/LogIn.tsx","./src/pages/Policy.tsx","./src/pages/Profile.tsx","./src/pages/ResetPassword.tsx","./src/pages/SignUp.tsx","./src/pages/UserPersonalInfo.tsx","./src/providers/AuthProvider.tsx","./src/providers/PHProvider.tsx","./src/providers/theme-provider.tsx","./src/services/authServices.ts","./src/services/drugServices.ts","./src/services/usersServices.ts","./src/types/types.ts"],"version":"5.8.3"}
+{"root":["./src/App.tsx","./src/constants.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/DevAlert.tsx","./src/components/ErrorBoundry.tsx","./src/components/ErrorElement.tsx","./src/components/FormFields.tsx","./src/components/Hero.tsx","./src/components/HydrateFallBack.tsx","./src/components/NotFound.tsx","./src/components/PrivateRoute.tsx","./src/components/ScrollToTopButton.tsx","./src/components/SpinnerOverlay.tsx","./src/components/WhatsAppButton.tsx","./src/components/MainLayout/ConsentSheet.tsx","./src/components/MainLayout/FixedUI.tsx","./src/components/MainLayout/Footer.tsx","./src/components/MainLayout/Header.tsx","./src/components/MainLayout/Logo.tsx","./src/components/MainLayout/MobileAuthBtns.tsx","./src/components/MainLayout/MobileMenu.tsx","./src/components/MainLayout/ProfilePic.tsx","./src/components/MainLayout/SigninAndLogoutButton.tsx","./src/components/MainLayout/app-sidebar.tsx","./src/components/MainLayout/mode-toggle.tsx","./src/components/auth/FaceBookOAuth.tsx","./src/components/auth/GoogleOAuth.tsx","./src/components/drugInfo/DrugInfoAccordion.tsx","./src/components/drugInfo/DrugPropertyDescription.tsx","./src/components/drugInfo/SearchDrugInfo.tsx","./src/components/drugInfo/drugCard.tsx","./src/components/drugList/InfinitScroll.tsx","./src/components/drugList/ListItem.tsx","./src/components/drugList/SelectRoute.tsx","./src/components/tellUsMore/SelectWithOther.tsx","./src/components/ui/accordion.tsx","./src/components/ui/alert.tsx","./src/components/ui/autocomplete.tsx","./src/components/ui/avatar.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/command.tsx","./src/components/ui/dialog.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/popover.tsx","./src/components/ui/select.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/sidebar.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/sonner.tsx","./src/components/ui/tooltip.tsx","./src/hooks/authContext.tsx","./src/hooks/use-mobile.ts","./src/hooks/useAuth.tsx","./src/hooks/useGetUser.ts","./src/hooks/useInfiniteScroll.ts","./src/hooks/useTheme.ts","./src/layouts/MainLayout.tsx","./src/lib/api.ts","./src/lib/firebase.ts","./src/lib/firebaseAuth.ts","./src/lib/firestore.ts","./src/lib/indexedDB.ts","./src/lib/queryQlient.ts","./src/lib/router.tsx","./src/lib/schemas.ts","./src/lib/utils.ts","./src/pages/DrugInfo.tsx","./src/pages/DrugList.tsx","./src/pages/GlobalError.tsx","./src/pages/Home.tsx","./src/pages/LogIn.tsx","./src/pages/Policy.tsx","./src/pages/Profile.tsx","./src/pages/ResetPassword.tsx","./src/pages/SignUp.tsx","./src/pages/UserPersonalInfo.tsx","./src/providers/AuthProvider.tsx","./src/providers/PHProvider.tsx","./src/providers/theme-provider.tsx","./src/services/authServices.ts","./src/services/drugServices.ts","./src/services/usersServices.ts","./src/types/types.ts"],"version":"5.8.3"}
